Bookweave v1.1: Collections
Adding the concept of Book collections. There are two new types of
linked transactions used by bookweave: book and collection. A book
belongs to a single collection. The collection is mainyl used to define
what are the metadata tags for the associated books. A more
complete description about them is given in the README. This new
version is not backwards compatible. You may still use v1.0 to view old
transactions.

Also separating the book publisher from the distributed app.
You are not allowed anymore to publish the books through the web,
instead you need to use the Python publisher. The app is only for
visualization purposes, and it uses Arweave's GraphQL API. For this
reason, there is a major cleanup in the distributed app.